Chore task, for issue 1701.
The issue is that the original code is not idiomatic rust, since it's
directly indexing acc via acc[pos]. However, looking at
merge_key_value_vecs, it already prefers the right value, so I _think_
we can just have the last two lines and get the same result.

Including the refactored code as a comment so we can have a discussion
on the PR, we may be doing the merge twice for cache line performance
reasons.